Improved codebase in aprgaudio module

+++ AprgAudio/AprgAudio/tst/AudioTests/Audio_unit.cpp
@@ -38,16 +38,14 @@ TEST(AiffLoadingTests, Stereo_8bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_8bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_8bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_8bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_8bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_8bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_16bit_44100) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\aiff_stereo_16bit_44100.aif)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -56,16 +54,14 @@ TEST(AiffLoadingTests, Stereo_16bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_16bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_16bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_16bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_16bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_24bit_44100) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\aiff_stereo_24bit_44100.aif)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -74,16 +70,14 @@ TEST(AiffLoadingTests, Stereo_24bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_24bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_24bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_24bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_24bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_24bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_8bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\aiff_stereo_8bit_48000.aif)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -92,16 +86,14 @@ TEST(AiffLoadingTests, Stereo_8bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_8bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_8bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_8bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_8bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_8bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_16bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\aiff_stereo_16bit_48000.aif)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -110,16 +102,14 @@ TEST(AiffLoadingTests, Stereo_16bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_16bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_16bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_16bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_16bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_24bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\aiff_stereo_24bit_48000.aif)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -128,16 +118,14 @@ TEST(AiffLoadingTests, Stereo_24bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_24bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_24bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_24bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_24bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_24bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WritingTest, WriteFromCopiedSampleBuffer) {
     Audio<float> audioFile1, audioFile2;
-
     bool isLoadingSuccessful =
         audioFile1.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_stereo_16bit_44100.wav)");
     ASSERT_TRUE(isLoadingSuccessful);
@@ -157,16 +145,14 @@ TEST(WavLoadingTests, Stereo_8bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_44100::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_8bit_44100::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_8bit_44100::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_16bit_44100) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_stereo_16bit_44100.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -175,16 +161,14 @@ TEST(WavLoadingTests, Stereo_16bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_44100::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_16bit_44100::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_16bit_44100::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_24bit_44100) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_stereo_24bit_44100.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -193,16 +177,14 @@ TEST(WavLoadingTests, Stereo_24bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_44100::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_24bit_44100::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_24bit_44100::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Mono_16bit_44100) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_mono_16bit_44100.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -211,16 +193,14 @@ TEST(WavLoadingTests, Mono_16bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_mono_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_mono_16bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < wav_mono_16bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_mono_16bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_mono_16bit_44100::testBuffer[i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_8bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_stereo_8bit_48000.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -229,16 +209,14 @@ TEST(WavLoadingTests, Stereo_8bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_48000::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_8bit_48000::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_8bit_48000::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_16bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_stereo_16bit_48000.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -247,16 +225,14 @@ TEST(WavLoadingTests, Stereo_16bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_48000::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_16bit_48000::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_16bit_48000::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_24bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_stereo_24bit_48000.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -265,16 +241,14 @@ TEST(WavLoadingTests, Stereo_24bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_48000::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_24bit_48000::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_24bit_48000::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Mono_16bit_48000) {
-    Audio<double> audioFile;
-    bool isLoadingSuccessful =
+    Audio<double> audioFile;    bool isLoadingSuccessful =
         audioFile.load(APRG_DIR R"(\AprgAudio\FilesForTests\ActualAudioFiles\wav_mono_16bit_48000.wav)");
 
     ASSERT_TRUE(isLoadingSuccessful);
@@ -283,33 +257,28 @@ TEST(WavLoadingTests, Mono_16bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_mono_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_mono_16bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < wav_mono_16bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_mono_16bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_mono_16bit_48000::testBuffer[i]));
         }
-    }
-}
+    }}
 
 bool writeTest(int numChannels, int sampleRate, int bitDepth, AudioFormat format) {
     float sampleRateAsFloat = (float)sampleRate;
-
     Audio<float> audioFile;
 
     audioFile.setAudioBufferSize(numChannels, sampleRate * 4);
 
-    for (unsigned int i = 0; i < static_cast<unsigned int>(audioFile.getNumSamplesPerChannel()); i++) {
+    for (int i = 0; i < static_cast<int>(audioFile.getNumSamplesPerChannel()); i++) {
         float sample = static_cast<float>(sin(2.0 * getPi() * i / sampleRateAsFloat * 440.0));
 
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
-            audioFile.samples[k][i] = sample * 0.5;
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) audioFile.samples[k][i] = sample * 0.5;
     }
 
-    audioFile.setSampleRate(sampleRate);
-    audioFile.setBitDepth(bitDepth);
+    audioFile.setSampleRate(sampleRate);    audioFile.setBitDepth(bitDepth);
 
     std::string numChannelsAsString = numChannels == 1 ? "mono" : "stereo";
-    std::string bitDepthAsString = std::to_string(bitDepth);
-    std::string sampleRateAsString = std::to_string(sampleRate);
+    std::string bitDepthAsString = std::to_string(bitDepth);    std::string sampleRateAsString = std::to_string(sampleRate);
 
     if (format == AudioFormat::Wave) {
         return audioFile.save(
